removed dual layer capability from foreground placement randomizer
