+++ SynthDet/Assets/Scenes/SynthDet.unity
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:26424c3a508656ea81a61ba91b040412eda4b7e1249e51b8ce1ae3b2467150b3
-size 100812
+oid sha256:fd7237002c8de88742ab5b711433a4520f3673d1c705f916c36ac66bf18eb375
+size 100387
