--- SynthDet/Assets/Scripts/Randomizers/DualLayerForegroundObjectPlacementRandomizer.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: a9dec5290d79747aabc09741e1f2bec6
-timeCreated: 1600838786
