+++ SynthDet/Assets/Scripts/Randomizers/ForegroundObjectPlacementRandomizer.cs
@@ -1,0 +1,81 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using UnityEngine;
+using UnityEngine.Perception.Randomization.Parameters;
+using UnityEngine.Perception.Randomization.Randomizers;
+using UnityEngine.Perception.Randomization.Randomizers.Utilities;
+using UnityEngine.Perception.Randomization.Samplers;
+
+namespace SynthDet.Randomizers
+{
+    /// <summary>
+    /// Creates a 2D layer of evenly spaced GameObjects from a given list of prefabs
+    /// </summary>
+    [Serializable]
+    [AddRandomizerMenu("Perception/My Foreground Object Placement Randomizer")]
+    public class ForegroundObjectPlacementRandomizer : Randomizer
+    {
+        List<GameObject> m_SpawnedObjects;
+        public int maxObjectCount = 100;
+        
+        public GameObjectParameter prefabs;
+
+        public float depth = 3f;
+        public FloatParameter separationDistance = new FloatParameter { value = new UniformSampler(0.7f, 1.2f) };
+        public Vector2 placementArea = new Vector2(5f, 5f);
+
+        GameObject m_Container;
+        GameObjectOneWayCache m_GameObjectOneWayCache;
+
+        int m_SpawnedCount;
+
+        protected override void OnAwake()
+        {
+            m_SpawnedObjects = new List<GameObject>();
+            m_Container = new GameObject("Foreground Objects");
+            var transform = scenario.transform;
+            m_Container.transform.parent = transform;
+            m_GameObjectOneWayCache = new GameObjectOneWayCache(m_Container.transform, prefabs.categories.Select(element => element.Item1).ToArray());
+        }
+    
+        /// <summary>
+        /// Generates a foreground layer of objects at the start of each scenario iteration
+        /// </summary>
+        protected override void OnIterationStart()
+        {
+            m_SpawnedObjects.Clear();
+            PlaceObjects();
+        }
+
+        void PlaceObjects()
+        {
+            m_SpawnedCount = 0;
+            
+            var seed = SamplerState.NextRandomState();
+            var placementSamples = PoissonDiskSampling.GenerateSamples(
+                placementArea.x, placementArea.y, separationDistance.Sample(), seed);
+            var offset = new Vector3(placementArea.x, placementArea.y, 0) * -0.5f;
+        
+            foreach (var sample in placementSamples)
+            {
+                var instance = m_GameObjectOneWayCache.GetOrInstantiate(prefabs.Sample());
+                instance.transform.position = new Vector3(sample.x, sample.y, depth) + offset;
+                m_SpawnedObjects.Add(instance);
+
+                if (++m_SpawnedCount == maxObjectCount)
+                    break;
+            }
+            placementSamples.Dispose();
+        }
+
+        /// <summary>
+        /// Deletes generated foreground objects after each scenario iteration is complete
+        /// </summary>
+        protected override void OnIterationEnd()
+        {
+            m_GameObjectOneWayCache.ResetAllObjects();
+        }
+    }
+}
+
