+++ SynthDet/Assets/Scripts/Randomizers/ForegroundObjectPlacementRandomizer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f1bfdc96f354d4356a5d5f162dbf3528
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
